\name{ynlb-package}
\alias{ynlb-package}
\alias{ynlb}
\docType{package}
\title{
\packageTitle{ynlb}
}
\description{
\packageDescription{ynlb}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ynlb}
\packageIndices{ynlb}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ynlb}

Maintainer: \packageMaintainer{ynlb}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
